+++ vtr-ui/src/app/components/ui/ui-brightness-slider/ui-brightness-slider.component.ts
@@ -12,7 +12,7 @@ export class UiBrightnessSliderComponent implements OnInit, AfterContentChecked 
 
 	@Input() enableSlider;
 	@Input() lightingBrightness: any;
-
+	@Input() ariaLabelValue = 'Brightness';
 	@Input() value = 1; // initial slider value
 	@Input() minValue = 1; // slider minimum end value
 	@Input() maxValue = 4; // slider maximum end value
@@ -46,8 +46,21 @@ export class UiBrightnessSliderComponent implements OnInit, AfterContentChecked 
 			floor: Number(this.minValue), // min value
 			ceil: Number(this.maxValue), // max value
 			step: Number(this.step), // value to change on each slide, default is 1
-			ariaLabel: 'Brightness'
+			ariaLabel: this.ariaLabelValue
 		};
+		document.body.addEventListener('mousedown', () => {
+			const elem = document.getElementsByClassName('brightness_slider')[0] as HTMLElement;
+			elem.setAttribute('role', '');
+			elem.setAttribute('role', 'slider');
+		});
+ 
+		document.body.addEventListener('keydown', (event) => {
+			if (event.keyCode === 9) {
+				const elem = document.getElementsByClassName('brightness_slider')[0] as HTMLElement;
+				elem.setAttribute('role', '');
+				elem.setAttribute('role', 'radiogroup');
+			}
+		});
 	}
 
 	public dragEnd($event: any) {