+++ vtr-ui/src/app/components/ui/ui-custom-radio/ui-custom-radio.ts
@@ -1,9 +1,26 @@
-import { ElementRef, Input, ViewChild } from '@angular/core';
+import { ElementRef, Input, ViewChild, AfterViewInit, OnInit, OnChanges } from '@angular/core';
 import { LoggerService } from 'src/app/services/logger/logger.service';
 import { MetricService } from 'src/app/services/metric/metric.service';
 import { KeyCode } from 'src/app/enums/key-code.enum';
 
-export class UICustomRadio {
+export class UICustomRadio implements OnInit, AfterViewInit, OnChanges {
+	static readonly ROLE_RADIO_GROUP = 'radiogroup';
+	static readonly ROLE = 'role';
+	static readonly DEPTH = 5;
+	static readonly UN_DEFINED = undefined;
+	static readonly TAB_INDEX = 'tabIndex';
+	static readonly ARIA_CHECKED = 'aria-checked';
+	// static readonly ARIA_DISABLED = 'aria-disabled';
+	static readonly NATIVE_ELEMENT = 'nativeElement';
+
+	constructor(private logger: LoggerService, public metrics: MetricService) { }
+	// once radio button is visible then execute logic
+	@ViewChild('radioButton', { static: false }) set content(element: ElementRef) {
+		if (element) {
+			this.radioButton = element;
+			this.setRadioButtons();
+		}
+	}
 	@Input() group: string;
 	@Input() label: string;
 	@Input() value: string;
@@ -15,20 +32,35 @@ export class UICustomRadio {
 	lastRadioButton: any;
 	radioButtons: Array<any> = [];
 	selectedRadioButton: any;
-	noRadioButtonSelected: boolean;
+	noRadioButtonSelected = false;
 	private radioButton: ElementRef<HTMLElement>;
 	radioLabel = 'radio.';
 
-	constructor(private logger: LoggerService,
-		public metrics: MetricService) { }
 
 	onChange(event) { }
-	// once radio button is visible then execute logic
-	@ViewChild('radioButton', { static: false }) set content(element: ElementRef) {
-		if (element) {
-			this.radioButton = element;
-			this.setRadioButtons();
-		}
+
+	ngOnInit() {
+		this.setRadioButtons(); // Set up radio buttons first , last etc and if none selected,set tabindex to first element
+	}
+
+	ngAfterViewInit(): void {
+		this.setRadioButtons(); // Set up radio buttons first , last etc and if none selected,set tabindex to first element
+	}
+
+	ngOnChanges(changes: any) {
+		// if (changes && changes.checked && !changes.checked.firstChange) {
+		// 	const elementDiv = document.getElementById('div' + this.radioId);
+		// 	if (elementDiv) {
+		// 		if (!this.checked) {
+		// 			elementDiv.setAttribute(UICustomRadio., 'false');
+		// 			elementDiv.tabIndex = -1;
+		// 		} else {
+		// 			elementDiv.setAttribute(UICustomRadio., 'true');
+		// 			elementDiv.tabIndex = 0;
+		// 			elementDiv.focus();
+		// 		}
+		// 	}
+		// }
 	}
 
 	changeRadioOnKeyPress($event, radio: HTMLInputElement) {
@@ -59,6 +91,7 @@ export class UICustomRadio {
 				// this.setChecked(this.radioButton.nativeElement, true);
 				$event.stopPropagation();
 				$event.preventDefault();
+				this.radioButton.nativeElement.focus();
 				break;
 			case KeyCode.UP:
 				this.setCheckedToPreviousItem(this.radioButton);
@@ -115,11 +148,11 @@ export class UICustomRadio {
 	private setRadioAriaChecked(currentRadioButton) {
 		try {
 			this.radioButtons.forEach(radioButton => {
-				radioButton.removeAttribute('aria-checked');
-				radioButton.setAttribute('aria-checked', 'false');
+				radioButton.removeAttribute(UICustomRadio.ARIA_CHECKED);
+				radioButton.setAttribute(UICustomRadio.ARIA_CHECKED, 'false');
 			});
-			currentRadioButton.removeAttribute('aria-checked');
-			currentRadioButton.setAttribute('aria-checked', 'true');
+			currentRadioButton.removeAttribute(UICustomRadio.ARIA_CHECKED);
+			currentRadioButton.setAttribute(UICustomRadio.ARIA_CHECKED, 'true');
 		} catch (error) {
 			this.logger.exception('setRadioAriaChecked error occurred ::', error);
 		}
@@ -130,7 +163,8 @@ export class UICustomRadio {
 			this.radioButtons.forEach(radioButton => {
 				radioButton.tabIndex = -1; // the unchecked item should also be tabbable
 			});
-			if (currentRadioButton !== undefined && currentRadioButton.tabIndex && currentRadioButton.tabIndex !== 0) {
+			// currentRadioButton[this.TABINDEX] !== UICustomRadio.UN_DEFINED && currentRadioButton.tabIndex !== 0
+			if (currentRadioButton !== UICustomRadio.UN_DEFINED) {
 				currentRadioButton.tabIndex = 0; // tabitem need not be set to 1 unnecessarly
 			}
 
@@ -158,18 +192,17 @@ export class UICustomRadio {
 
 	private setCheckedToNextItem(currentItem) {
 		try {
-			if (currentItem) {
-				let index;
+			let index;
 
-				if (currentItem.nativeElement === this.lastRadioButton) {
-					this.setChecked(this.firstRadioButton, false);
-				} else {
-					index = this.radioButtons.indexOf(currentItem.nativeElement);
-					this.setChecked(this.radioButtons[index + 1], false);
-				}
+			if (currentItem.nativeElement === this.lastRadioButton) {
+				this.setChecked(this.firstRadioButton, false);
+			} else {
+				index = this.radioButtons.indexOf(currentItem.nativeElement);
+				this.setChecked(this.radioButtons[index + 1], false);
 			}
+
 		} catch (error) {
-			this.logger.exception('UICustomRadio.setCheckedToNextItem exception', error);
+			this.logger.exception('setRadioButtons error occurred ::', error);
 		}
 
 	}
@@ -177,44 +210,43 @@ export class UICustomRadio {
 	protected setRadioButtons() {
 		try {
 			const rbs = this.getRadioGroup();
-
 			this.radioButtons = [];
-			rbs.forEach(radioButton => {
-				this.radioButtons.push(radioButton);
-				if (this.firstRadioButton === undefined || !this.firstRadioButton) {
-					this.firstRadioButton = radioButton;
-				}
-				this.lastRadioButton = radioButton;
+			if (rbs !== UICustomRadio.UN_DEFINED && rbs && rbs.length > 0) {
+				rbs.forEach(radioButton => {
+					this.radioButtons.push(radioButton);
+					if (this.firstRadioButton === UICustomRadio.UN_DEFINED || !this.firstRadioButton) {
+						this.firstRadioButton = radioButton;
+					}
+					this.lastRadioButton = radioButton;
 
-				if (radioButton.getAttribute('aria-checked') === 'true') {
-					this.selectedRadioButton = radioButton;
-					this.noRadioButtonSelected = false;
+					if (radioButton.getAttribute(UICustomRadio.ARIA_CHECKED) === 'true') {
+						this.selectedRadioButton = radioButton;
+						this.noRadioButtonSelected = false;
+					}
+				});
+				// focus on first non disabled element if not selected any radio items
+				// || this.firstRadioButton.tabIndex !== 0
+				if (this.noRadioButtonSelected && this.firstRadioButton !== UICustomRadio.UN_DEFINED) {
+					this.setRadioTabIndex(this.getNativeElement(this.firstRadioButton));
 				}
-			});
+				// || this.selectedRadioButton.tabIndex !== 0
+				else if (!this.noRadioButtonSelected && this.selectedRadioButton !== UICustomRadio.UN_DEFINED) {
+					this.setRadioTabIndex(this.getNativeElement(this.selectedRadioButton));
+				}
+			}
 
-			// focus on first non disabled element if not selected any radio items
-			if (this.noRadioButtonSelected && this.firstRadioButton !== undefined
-				&& (this.firstRadioButton.tabIndex || this.firstRadioButton.tabIndex !== 0)
-			) {
-				this.setRadioTabIndex(this.firstRadioButton.nativeElement);
-			}
-			else if (!this.noRadioButtonSelected && this.selectedRadioButton !== undefined
-				&& (this.selectedRadioButton.tabIndex || this.selectedRadioButton.tabIndex !== 0)) {
-				this.setRadioTabIndex(this.selectedRadioButton.nativeElement);
-			}
 		} catch (error) {
-			this.logger.exception('UICustomRadio.setRadioButtons exception', error);
+			this.logger.exception('setRadioButtons error occurred ::', error);
 		}
 	}
 
 	private getRadioGroup() {
 
-		// commented to remove the dependency from radiogroup role of parent this component
 		try {
+			// commented to remove the dependency from radiogroup role of parent this component
 
-			const radioElement = (this.radioButton.nativeElement ? this.radioButton.nativeElement : this.radioButton);
-			if (!this.radioGroup && radioElement) {
-				this.radioGroup = this.getParentRadioGroup(radioElement, 10);
+			this.radioGroup = this.getParentRadioGroup(this.getNativeElement(this.radioButton), UICustomRadio.DEPTH);
+			if (this.radioGroup !== UICustomRadio.UN_DEFINED) {
 				// search by radio class and aria-disabled
 				const query = `[class*=${this.group}][aria-disabled=false]`;
 				// search by only role and aria-disabled
@@ -222,23 +254,35 @@ export class UICustomRadio {
 				return this.radioGroup.querySelectorAll(query);
 				// return Array.from(this.radioGroup.querySelectorAll(query));
 			}
+
 		} catch (error) {
 			this.logger.exception('UICustomRadio.getRadioGroup exception', error);
 		}
 
-
 	}
+
+	private getNativeElement(element) {
+		try {
+			if (element !== UICustomRadio.UN_DEFINED) {
+				element = element && element[UICustomRadio.NATIVE_ELEMENT] ? element.nativeElement : element;
+			}
+			return element;
+		}
+		catch (error) {
+			this.logger.exception('UICustomRadio.getElement exception', error);
+		}
+	}
+
 
 	private getParentRadioGroup(element, topUpLevel: number) {
 		try {
-			const roleRadioGroup = 'radiogroup';
-			const role = 'role';
 
-			if (element && element.getAttribute(role) === roleRadioGroup) {
+
+			if (element && element.getAttribute(UICustomRadio.ROLE) === UICustomRadio.ROLE_RADIO_GROUP) {
 				return element;
 			}
 			else if (element && element.parentElement
-				&& element.getAttribute(role) !== roleRadioGroup && topUpLevel > 0) {
+				&& element.getAttribute(UICustomRadio.ROLE) !== UICustomRadio.ROLE_RADIO_GROUP && topUpLevel > 0) {
 				return this.getParentRadioGroup(element.parentElement, --topUpLevel);
 			}
 			else {
@@ -250,5 +294,6 @@ export class UICustomRadio {
 
 		}
 	}
+
 
 }