+++ vtr-ui/src/app/components/widgets/power-smart-settings/power-smart-settings.component.spec.ts
@@ -171,7 +171,7 @@ describe("Component: PowerSmartSetting", () => {
 		let spy = spyOn(powerService, 'getITSModeForICIdeapad').and.returnValue(Promise.resolve(response))
 		component.initPowerSmartSettingsForIdeaPad()
 		expect(spy).toHaveBeenCalled()
-	}));	
+	}));
 
 	it('should call initPowerSmartSettingsForIdeaPad - available is present', async(() => {
 		let response = {
@@ -318,7 +318,7 @@ describe("Component: PowerSmartSetting", () => {
 		// spyOn(powerService, 'getAMTCapability').and.returnValue(Promise.resolve(true))
 		component.initPowerSmartSettingsForThinkPad()
 		done()
-		expect(component.cache.showIC).toEqual(component.showIC)		
+		expect(component.cache.showIC).toEqual(component.showIC)
 	}));
 
 	it('should call dytc6GetStatus', () => {
@@ -371,18 +371,6 @@ describe("Component: PowerSmartSetting", () => {
 		powerService.isShellAvailable = true
 		component['getEMDriverStatus']()
 		expect(component['getEMDriverStatus']).toBeTruthy()
-	});
-
-	it('should call isShowIntelligentCoolingModes', () => {
-		component.legacyManualModeCapability = false
-		component.isShowIntelligentCoolingModes()
-		expect(component.isShowIntelligentCoolingModes).toBeTruthy()
-	});
-
-	it('should call isShowIntelligentCoolingModes -else', () => {
-		component.legacyManualModeCapability = true
-		component.isShowIntelligentCoolingModes()
-		expect(component.isShowIntelligentCoolingModes).toBeTruthy()
 	});
 
 	it('should call getDYTCRevision - catch block', () => {