+++ vtr-ui/src/app/modules/app-routing.module.ts
@@ -84,6 +84,11 @@ const routes: Routes = [
 		canActivate: [ProtocolGuardService],
 		component: HomeComponent,
 		pathMatch: 'full'
+	},
+	{
+		path: '**',
+		canActivate: [ProtocolGuardService],
+		component: HomeComponent
 	}
 ];