+++ vtr-ui/src/app/services/smart-assist/smart-assist.service.ts
@@ -5,7 +5,6 @@ import {
 	FeatureStatus
 } from 'src/app/data-models/common/feature-status.model';
 import { VantageShellService } from '../vantage-shell/vantage-shell.service';
-import { DisplayService } from 'src/app/services/display/display.service';
 import { AntiTheftResponse } from 'src/app/data-models/antiTheft/antiTheft.model';
 
 @Injectable({
@@ -20,11 +19,6 @@ export class SmartAssistService {
 	private superResolution;
 	private antiTheft;
 	public windows;
-	private DeviceInformation: any;
-	private DeviceClass: any;
-	private Front: any;
-	private mediaCapture: any;
-	private Capture: any;
 	private hsaIntelligentSecurity;
 	private hpdAdvancedSettings = {
 		zeroTouchLoginAdvanced: false,
@@ -35,7 +29,7 @@ export class SmartAssistService {
 	public isAPSavailable = false;
 	private shellVersion: string;
 
-	constructor(shellService: VantageShellService, private displayService: DisplayService) {
+	constructor(shellService: VantageShellService) {
 		this.intelligentSensing = shellService.getIntelligentSensing();
 		this.intelligentMedia = shellService.getIntelligentMedia();
 		this.activeProtectionSystem = shellService.getActiveProtectionSystem(); // getting APS Object from //vantage-shell.service
@@ -45,12 +39,6 @@ export class SmartAssistService {
 		// this.shellVersion = shellService.getShellVersion();
 		this.antiTheft = shellService.getAntiTheft();
 		this.windows = shellService.getWindows();
-		if (this.windows) {
-			this.DeviceInformation = this.windows.Devices.Enumeration.DeviceInformation;
-			this.DeviceClass = this.windows.Devices.Enumeration.DeviceClass;
-			this.Front = this.windows.Devices.Enumeration.Panel.front;
-			this.Capture = this.windows.Media.Capture;
-		}
 		this.activeProtectionSystem ? this.isAPSavailable = true : this.isAPSavailable = false;
 		if (this.intelligentSensing && this.intelligentMedia && this.lenovoVoice && this.superResolution) {
 			this.isShellAvailable = true;
@@ -359,11 +347,11 @@ export class SmartAssistService {
 	//region Intelligent Sensting (anti theft) section
 
 	public async getAntiTheftStatus(): Promise<AntiTheftResponse> {
-		const antiTheftDate = { available: false, status: false, isSupportPhoto: false, photoAddress: "", cameraPrivacyState: false, authorizedAccessState: false, alarmOften: 0, photoNumber: 5 };
+		const antiTheftDate = { available: false, status: false, isSupportPhoto: false, photoAddress: "", cameraPrivacyState: true, authorizedAccessState: true, alarmOften: 0, photoNumber: 5 };
 		try {
 			if (this.isShellAvailable) {
-				const data = await Promise.all([this.getCameraAuthorizedAccessState(), this.antiTheft.getMotionAlertSetting(), this.getCameraPrivacyState()]);
-				const obj = JSON.parse(data[1]);
+				const data = this.antiTheft.getMotionAlertSetting();
+				const obj = JSON.parse(data);
 				if (obj && obj.errorCode === 0) {
 					antiTheftDate.available = obj.available;
 					antiTheftDate.status = obj.enabled;
@@ -372,51 +360,11 @@ export class SmartAssistService {
 					antiTheftDate.alarmOften = obj.alarmDuration;
 					antiTheftDate.photoNumber = obj.photoNumber;
 				}
-				antiTheftDate.cameraPrivacyState = !data[2];
-				antiTheftDate.authorizedAccessState = data[0];
 				return Promise.resolve(antiTheftDate);
 			}
 		} catch (error) {
 			return Promise.resolve(antiTheftDate);
 		}
-	}
-
-	public getCameraAuthorizedAccessState() {
-		let deviceInfo = null;
-		return this.DeviceInformation.findAllAsync(this.DeviceClass.videoCapture)
-			.then((devices: any) => {
-				devices.forEach((cameraDeviceInfo: any) => {
-					if (cameraDeviceInfo.enclosureLocation != null && cameraDeviceInfo.enclosureLocation.panel === this.Front) {
-						deviceInfo = cameraDeviceInfo;
-					}
-				});
-				if (deviceInfo && devices.length > 0) {
-					deviceInfo = devices.getAt(0);
-					this.mediaCapture = new this.Capture.MediaCapture();
-					const settings = new this.Capture.MediaCaptureInitializationSettings();
-					settings.videoDeviceId = deviceInfo.id;
-					settings.streamingCaptureMode = 2;
-					settings.photoCaptureSource = 0;
-					return this.mediaCapture.initializeAsync(settings).then(() => {
-						return Promise.resolve(true);
-					}, (error: any) => {
-						if (error && error.number === -2147024891) {
-							return Promise.resolve(false);
-						}
-					});
-				}
-				return Promise.resolve(false);
-			}, (error: any) => {
-				return Promise.resolve(false);
-			});
-	}
-
-	public async getCameraPrivacyState(): Promise<boolean> {
-		if (this.displayService.isShellAvailable && this.isShellAvailable) {
-			const featureStatus = await this.displayService.getCameraPrivacyModeState();
-			return featureStatus.status;
-		}
-		return undefined;
 	}
 
 	public setAntiTheftStatus(value: boolean): Promise<boolean> {